Introduce Modified trait to create a modified-state tree graph
Project owns objects that can be modified and saved to disk, like the
font's metadata object, FontInfo.

To propagate modified property changes up to Project itself with ease,
we introduce a trait that binds the modified properties of parent and
child and only sets the parent's modified property to true if the child
property is also true. This means that if we change e.g. a single glyph,
the Project object will become modified as well. Then we can call save()
on Project, and it will save anything that has the modified flag on.